pyproject.toml
[tool.poetry]
name = "knitting-pattern-maker"
version = "0.1.0"
description = ""
authors = ["Asgeir Melling <asgmel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
